Plataforma_de_Vendas/requirements.txt
Django==4.2.4
djangorestframework==3.13.1
drf_yasg==1.20.0
nanoid